Move logging extensions out of configuration
